% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distance_elevation_df.R
\name{create_distance_elevation_df}
\alias{create_distance_elevation_df}
\title{Create a data frame with distance elevation pairs based on track points}
\usage{
create_distance_elevation_df(track_points_sf)
}
\arguments{
\item{track_points_sf}{An sf object, typically generated with `read_track_points()`}
}
\value{
A data frame with two columns, distance and elevation.
}
\description{
Create a data frame with distance elevation pairs based on track points
}
