% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_miniature.R
\name{create_miniature}
\alias{create_miniature}
\title{Create miniature preview of the track}
\usage{
create_miniature(
  track_points_sf,
  tracks_sf,
  title = "GPX track",
  font_family = "sans"
)
}
\arguments{
\item{track_points_sf}{}

\item{tracks_sf}{}
}
\value{

}
\description{
Create miniature preview of the track
}
