% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qr.R
\name{create_qr}
\alias{create_qr}
\title{Create qr code from URL}
\usage{
create_qr(url)
}
\arguments{
\item{url}{A url to be transformed into a QR code}
}
\value{
A ggplot object
}
\description{
Create qr code from URL
}
\examples{

create_qr("https://archiviomemoria.ecomuseovalledeilaghi.it")
}
