% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tracks.R
\name{read_track_points}
\alias{read_track_points}
\title{Read track points}
\usage{
read_track_points(path)
}
\arguments{
\item{path}{Path to gpx file.}
}
\value{
An sf object
}
\description{
Reads either tracks points or route points as sf objects If tracks is empty, it reads routes in its stead.
}
