% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tracks.R
\name{read_tracks}
\alias{read_tracks}
\title{Read tracks}
\usage{
read_tracks(path)
}
\arguments{
\item{path}{Path to gpx file.}
}
\value{
An sf object
}
\description{
Reads either tracks or routes as sf objects. If tracks is empty, it reads routes in its stead.
}
